Avoid doing things on non-existing DOM element

+++ plugins/antispam/js/antispam.js
@@ -4,31 +4,33 @@
 Object.assign(dotclear.msg, dotclear.getData('antispam'));
 
 $(() => {
-  $('#filters-list').sortable({
-    cursor: 'move',
-  });
-  $('#filters-list tr')
-    .on('mouseenter', function () {
-      $(this).css({
-        cursor: 'move',
-      });
-    })
-    .on('mouseleave', function () {
-      $(this).css({
-        cursor: 'auto',
-      });
+  if ($('#filters-list').length) {
+    $('#filters-list').sortable({
+      cursor: 'move',
     });
-  $('#filters-list-form').on('submit', () => {
-    const order = [];
-    $('#filters-list tr td input.position').each(function () {
-      order.push(this.name.replace(/^f_order\[([^\]]+)\]$/, '$1'));
+    $('#filters-list tr')
+      .on('mouseenter', function () {
+        $(this).css({
+          cursor: 'move',
+        });
+      })
+      .on('mouseleave', function () {
+        $(this).css({
+          cursor: 'auto',
+        });
+      });
+    $('#filters-list-form').on('submit', () => {
+      const order = [];
+      $('#filters-list tr td input.position').each(function () {
+        order.push(this.name.replace(/^f_order\[([^\]]+)\]$/, '$1'));
+      });
+      $('input[name=filters_order]')[0].value = order.join(',');
+      return true;
     });
-    $('input[name=filters_order]')[0].value = order.join(',');
-    return true;
-  });
-  $('#filters-list tr td input.position').hide();
-  $('#filters-list tr td.handle').addClass('handler');
+    $('#filters-list tr td input.position').hide();
+    $('#filters-list tr td.handle').addClass('handler');
 
-  $('form input[type=submit][name=delete_all]').on('click', () => window.confirm(dotclear.msg.confirm_spam_delete));
-  dotclear.responsiveCellHeaders(document.querySelector('#filters-list-form table'), '#filters-list-form table', 1, true);
+    $('form input[type=submit][name=delete_all]').on('click', () => window.confirm(dotclear.msg.confirm_spam_delete));
+    dotclear.responsiveCellHeaders(document.querySelector('#filters-list-form table'), '#filters-list-form table', 1, true);
+  }
 });
